Added PoS caching

+++ Source/twitter-sent-hybrid/transformer/pos_transformer.py
@@ -1,9 +1,9 @@
 from sklearn.base import TransformerMixin, BaseEstimator
 from sklearn.feature_extraction import DictVectorizer
-from sklearn.feature_extraction.text import TfidfVectorizer
 from sklearn.preprocessing import normalize
 import nltk
 import utils.preprocessor_methods as t
+from storage import cache
 
 
 class POSTransformer(TransformerMixin, BaseEstimator):
@@ -14,16 +14,20 @@ class POSTransformer(TransformerMixin, BaseEstimator):
         self.pos_dict = dict.fromkeys(POSTransformer.classes, 0)
 
     def transform(self, raw_tweets, **transform_params):
-        occurences = []
-        for tweet in raw_tweets:
-            tweet = t.remove_all(tweet)
-            tokens = nltk.word_tokenize(tweet)
-            tagged = nltk.pos_tag(tokens)
-            tag_frequencies = self.pos_dict.copy()
-            for word, tag in tagged:
-                if tag in tag_frequencies:
-                    tag_frequencies[tag] += 1
-            occurences.append(tag_frequencies)
+        occurences = cache.get(str(raw_tweets), "pos_frequencies_")
+
+        if not occurences:
+            occurences = []
+            for tweet in raw_tweets:
+                tweet = t.remove_all(tweet)
+                tokens = nltk.word_tokenize(tweet)
+                tagged = nltk.pos_tag(tokens)
+                tag_frequencies = self.pos_dict.copy()
+                for word, tag in tagged:
+                    if tag in tag_frequencies:
+                        tag_frequencies[tag] += 1
+                occurences.append(tag_frequencies)
+            cache.save(str(raw_tweets), occurences, "pos_frequencies_")
         vectorized = self.vectorizer.transform(occurences)
         return normalize(vectorized) if self.normalize else vectorized
